Addressed some comments
