+++ backend/src/types/NotificationEnums.ts
@@ -1,0 +1,10 @@
+export enum NotificationType {
+    WEB = "web",
+    EMAIL = "email",
+}
+
+export enum NotificationStatus {
+    UNREAD = "unread",
+    READ = "read",
+    ACKNOWLEDGED = "acknowledged",
+}
